# Conflicts:
#	backend/backend/aceofclubs/fixtures/Fixtures.json
#	frontend/src/app/services/user.service.ts

+++ frontend/src/app/services/user.service.ts
@@ -3,6 +3,10 @@ import {HttpClient} from "@angular/common/http";
 import {BehaviorSubject, Observable} from "rxjs";
 import {Router} from "@angular/router";
 import {JwtHelperService} from '@auth0/angular-jwt';
+import {Group} from "./group.service";
+import {UserGroupService} from "./user-group.service";
+
+
 
 
 export interface User {
@@ -24,7 +28,8 @@ export class UserService {
 
   readonly accessTokenLocalStorageKey = 'access_token';
   isLoggedIn = new BehaviorSubject(false);
-  currentUserPK: number;
+  currentUserPK:number;
+  currentUserName:string;
 
 
   constructor(private http: HttpClient, private router: Router, private jwtHelperService: JwtHelperService) {
@@ -42,8 +47,19 @@ export class UserService {
   }
 
   deleteUser(user: User): Observable<any> {
-    return this.http.delete('/api/users/' + user.pk + '/');
+    if (this.currentUserPK != 1) {
+      return this.http.delete('/api/users/' + user.pk + '/');
+    }
+    else {
+      alert('Users may only be deleted by Administrators!');
+      /** TO DO: SONST ACCOUNT DEACTIVATEN UND
+       * MIT HAS_PERMISSION ABFRAGEN OB DERJENIGE ADMIN IST!**/
+    }
   };
+
+  /*deactivateUser(user: User): Observable<any> {
+    return this.http.delete('/api/users/' + user.pk + '/');
+  };*/
 
   getUser(pk: number): Observable<User> {
     return this.http.get<User>('/api/users/' + pk + '/');
@@ -64,7 +80,7 @@ export class UserService {
         localStorage.setItem('currentUser', userData.username);
         localStorage.setItem('access_token', res.token);
         this.router.navigate(['event-list']);
-        this.cUID();
+        this.getCurrentUserId();
         alert('Logged in as: ' + localStorage.getItem('currentUser'))
       }, () => {
         alert('wrong username or password');
@@ -80,15 +96,25 @@ export class UserService {
   }
 
 
+  // Aktuell angemeldeten User mittels username ermitteln
   getCurrentUser(): Observable<User> {
     return this.http.get<User>('/api/users/?username=' + localStorage.getItem('currentUser'));
   }
-
-  cUID(): void {
-    this.getCurrentUser().subscribe((user) => {
+  // PK des aktuell angemeldeten User in Variable speichern
+  getCurrentUserId():void{
+    this.getCurrentUser().
+    subscribe((user)=>{
       this.currentUserPK = user[0].pk
     })
   }
+
+  currentUsername():void{
+    this.getCurrentUser().
+    subscribe((user)=>{
+      this.currentUserName = user[0].username
+    })
+  }
+
 
   hasPermission(permission: string): boolean {
     const token = localStorage.getItem(this.accessTokenLocalStorageKey);
